Move websocket hooks to dedicated folder for improved organisation
